moved angle-to-rot-matrix
